Yes, most has been changed, all thats done was adding the lic information and the CC info in the headers. Time to mark the code properly.
Code that was not developed by me, carries its orginal header with lic

+++ src/lightsensorfeature.cpp
@@ -27,8 +27,8 @@ LightSensorFeature::~LightSensorFeature()
 
 void LightSensorFeature::run()
 {
-    int pinRead;
-    int pinLastState = 1;
+    short pinRead;
+    short    pinLastState = 1;
     std::mutex mtx;
     std::unique_lock<std::mutex> lck(mtx);
     time_t nextTrip = 0;
@@ -53,7 +53,6 @@ void LightSensorFeature::run()
                     emit lightSensorStateChange(this, pinLastState);
                     nextTrip = time(nullptr) + 120;
                 }
-
         }
 
     }
